requirements.txt
beautifulsoup4==4.4.1
blinker==1.3
chardet==2.3.0
checkbox-support==0.22
cryptography==1.2.3
cx-Oracle==7.0.0
defusedxml==0.5.0
Django==2.1.1
django-allauth==0.22.0
django-filter==2.0.0
django-rest-auth==0.9.3
django-templated-mail==1.1.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
djoser==1.3.0
feedparser==5.1.3
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.0
Jinja2==2.8
lxml==3.5.0
Mako==1.0.3
Markdown==2.6.11
MarkupSafe==0.23
mysqlclient==1.3.13
oauthlib==1.0.3
padme==1.1.1
pexpect==4.0.1
Pillow==3.1.2
plainbox==0.25
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
PyICU==1.9.2
PyJWT==1.3.0
pyparsing==2.0.3
python3-openid==3.1.0
pytz==2018.5
pyxdg==0.25
reportlab==3.3.0
requests==2.9.1
requests-oauthlib==1.0.0
six==1.10.0
urllib3==1.13.1
XlsxWriter==0.7.3
